Refactor merge vhdx into seperate files
